requirements.txt
# GUI y Lógica Principal
FreeSimpleGUI
pandas
numpy
networkx
matplotlib
Pillow
reportlab
xhtml2pdf
cefpython3